pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["timedb", "timedb.db", "timedb.workflows"]

[tool.setuptools.package-data]
timedb = ["sql/*.sql"]

[project]
name = "timedb"
version = "0.1.1"
description = "timedb — opinionated schema & API for time series"
readme = "README.md"
requires-python = ">=3.8"

# runtime dependencies (adjust versions as you prefer)
dependencies = [
  "click>=8.0",
  "psycopg[binary]>=3.1",     # psycopg3 with binary wheels on macOS
  "python-dotenv>=0.21",      # load_dotenv used in helper scripts
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "pandas>=2.0.0",
  "sqlalchemy>=2.0.0",
]

# expose the CLI command 'timedb' which points to the Click group `cli` in timedb/cli.py
[project.scripts]
timedb = "timedb.cli:cli"